+++ pv_tp3_grupo1/src/App.jsx
@@ -1,35 +1,38 @@
+import React from 'react';
 import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
-import './App.css'
-
+import TaskList from './assets/components/TaskList.jsx'
+import TaskInput from './assets/components/TaskInput.jsx';
 function App() {
-  const [count, setCount] = useState(0)
+  const [tasks, setTasks] = useState([]); // lista de tareas
+
+  const agregarTarea = (texto) => {
+    const nuevaTarea = {
+      id: crypto.randomUUID(), // genera un ID único
+      text: texto,
+      completed: false
+    };
+    setTasks([...tasks, nuevaTarea]);
+  };
+
+  const onToggleComplete = (id) => {
+    const nuevasTareas = tasks.map((task, i) =>
+      i === id ? { ...task, completed: !task.completed } : task
+    );
+    setTasks(nuevasTareas);
+  };
+
 
   return (
-    <>
-      <div>
-        <a href="https://vite.dev" target="_blank">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
+    <div className='container'>
+      <div className='tareas'>
+        <h1>LISTA DE TAREAS</h1>
+        <TaskInput onAdd={agregarTarea}/>
+        <TaskList tasks={tasks} onToggleComplete={onToggleComplete} />
       </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.jsx</code> and save to test HMR
-        </p>
+      <div className='productos'>
+
       </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
-    </>
+    </div>
   )
 }
-
 export default App
